requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-crum==0.7.9
python-decouple==3.6
sqlparse==0.4.2
